🚚 move: templates 내부 컴포넌트 -> index.tsx
