--- app/src/components/templates/DashboardSkeleton/index.ts
@@ -1,1 +1,0 @@
-export * from './DashboardSkeleton';
