--- app/src/components/templates/LeaderBoard/index.ts
@@ -1,1 +1,0 @@
-export * from './LeaderBoard';
