+++ app/src/pages/EvalLogSearchPage/EvalLogSearchDetail.tsx
@@ -2,7 +2,7 @@ import { EvalLogEdge } from '@/__generated__/graphql';
 import { ApolloError } from '@apollo/client';
 import { Center, Skeleton, Text, VStack } from '@components/common';
 import { ApolloBadRequest } from '@components/elements/DashboardContentView/Error';
-import { Footer } from '@components/elements/Footer/Footer';
+import { Footer } from '@components/elements/Footer';
 import { useTheme } from '@emotion/react';
 import React from 'react';
 import { EvalLogList } from './EvalLogList';
